refactor: reorgnize theme package
