+++ packages/mask/src/plugins/PoolTogether/UI/Account.tsx
@@ -1,5 +1,5 @@
 import { RefreshIcon } from '@masknet/icons'
-import { DarkColor } from '@masknet/theme/constants'
+import { DarkColor } from '@masknet/theme/base'
 import { formatBalance } from '@masknet/web3-shared-evm'
 import { CircularProgress, Grid, Link, Typography } from '@mui/material'
 import { makeStyles } from '@masknet/theme'
