+++ packages/mask/src/plugins/PoolTogether/UI/CountdownView.tsx
@@ -3,7 +3,7 @@ import { Box, Grid, Typography } from '@mui/material'
 import { makeStyles, useStylesExtends } from '@masknet/theme'
 import { useInterval } from 'react-use'
 import { ONE_SECOND } from '../constants'
-import { DarkColor } from '@masknet/theme/constants'
+import { DarkColor } from '@masknet/theme/base'
 import intervalToDuration from 'date-fns/intervalToDuration'
 
 const useStyles = makeStyles()((theme) => ({
