+++ packages/theme/src/CSSVariables/Injector.tsx
@@ -1,6 +1,6 @@
 import { GlobalStyles, Theme, useTheme } from '@mui/material'
 import { useRef } from 'react'
-import { CSSVariableInjectorCSS } from './constants'
+import { CSSVariableInjectorCSS } from './applyToDOM'
 
 export function CSSVariableInjector(props: React.PropsWithChildren<{ useTheme?: () => Theme }>) {
     const { current: useConsistentTheme } = useRef(props.useTheme || useTheme)
