+++ packages/theme/src/Components/PhoneNumberField/index.tsx
@@ -1,5 +1,5 @@
 import { Box, InputBaseProps } from '@mui/material'
-import { makeStyles } from '../../makeStyles'
+import { makeStyles } from '../../UIHelper/makeStyles'
 import { MaskTextField } from '../TextField'
 import { ChangeEvent, ReactNode, useState } from 'react'
