+++ packages/theme/src/Components/Snackbar/index.tsx
@@ -18,9 +18,9 @@ import WarningIcon from '@mui/icons-material/Warning'
 import InfoIcon from '@mui/icons-material/Info'
 import DoneIcon from '@mui/icons-material/Done'
 import { LoadingIcon, RiskIcon } from '@masknet/icons'
-import { makeStyles } from '../../makeStyles'
+import { makeStyles } from '../../UIHelper/makeStyles'
 import { useStylesExtends } from '../../UIHelper/custom-ui-helper'
-import { MaskColorVar } from '../../constants'
+import { MaskColorVar } from '../../CSSVariables'
 
 export { SnackbarProvider, useSnackbar } from 'notistack'
 export type { VariantType, OptionsObject, SnackbarKey, SnackbarMessage } from 'notistack'
