+++ packages/theme/src/Theme/Provider.tsx
@@ -1,6 +1,6 @@
 import { CssBaseline, Theme, ThemeProvider } from '@mui/material'
 import { MaskIconPalette, MaskIconPaletteContext } from '@masknet/icons'
-import { CustomSnackbarProvider } from './Components/Snackbar'
+import { CustomSnackbarProvider } from '../Components/Snackbar'
 
 function compose(init: React.ReactNode, ...f: ((children: React.ReactNode) => JSX.Element)[]) {
     return f.reduceRight((prev, curr) => curr(prev), <>{init}</>)
