+++ packages/theme/src/Theme/component-changes.ts
@@ -1,21 +1,7 @@
 import type { PaletteMode, ThemeOptions } from '@mui/material'
-import type { Color } from './constants'
+import type { Color } from '../CSSVariables'
+import { css_var } from '../CSSVariables/createVars'
 
-function css_var<T extends Record<string, unknown>>(
-    unique_name: string,
-    keys: T,
-): { [key in keyof T]: string & ((defaultValue?: string) => string) } {
-    for (const k in keys) keys[k] = createVar(k) as any
-    return keys as any
-    function createVar(name: string) {
-        const val = '--' + unique_name + '-' + name
-        function use(defaultValue?: string): string {
-            return `var(${val}${typeof defaultValue === 'undefined' ? '' : ', ' + defaultValue})`
-        }
-        use.toString = () => val
-        return use
-    }
-}
 type Theme = ThemeOptions | ((mode: PaletteMode, colors: Color) => ThemeOptions)
 
 export const BaseLine: Theme = (mode, colors): ThemeOptions => ({
