+++ packages/theme/src/Theme/index.ts
@@ -1,0 +1,3 @@
+export * from './types'
+export { MaskThemeProvider } from './Provider'
+export { MaskDarkTheme, MaskLightTheme } from './theme'
