+++ packages/theme/src/Theme/types.ts
@@ -1,0 +1,5 @@
+export enum Appearance {
+    default = 'default',
+    light = 'light',
+    dark = 'dark',
+}
