+++ packages/theme/src/UIHelper/makeStyles.ts
@@ -1,5 +1,3 @@
 import { createMakeStyles } from 'tss-react'
 import { useTheme } from '@mui/material'
 export const { makeStyles } = createMakeStyles({ useTheme })
-export type { Css, Cx } from 'tss-react'
-export { keyframes } from 'tss-react'
