+++ packages/theme/src/UIHelper/useUserPrefersColorScheme.ts
@@ -1,0 +1,5 @@
+import { type PaletteMode, useMediaQuery } from '@mui/material'
+
+export function useSystemPreferencePalette(): PaletteMode {
+    return useMediaQuery('(prefers-color-scheme: dark)') ? 'dark' : 'light'
+}
