--- packages/theme/src/customization/types.ts
@@ -1,6 +1,0 @@
-type ClassNameMap<ClassKey extends string = string> = { [P in ClassKey]: string }
-
-export interface ComponentOverwriteConfig<Props extends { classes?: any }, Classes extends string> {
-    classes?: () => { classes: Partial<ClassNameMap<Classes>> }
-    props?: (props: Props) => Props
-}
