+++ packages/theme/src/entry-base.ts
@@ -1,0 +1,12 @@
+/// <reference path="./extended.d.ts" />
+
+// Some re-exports
+import tinyColor from 'tinycolor2'
+export const parseColor = tinyColor
+export { keyframes, type Css, type Cx, TssCacheProvider } from 'tss-react'
+
+//
+export * from './UIHelper'
+export * from './CSSVariables'
+export * from './Theme'
+export * from './ShadowRoot'
